+++ packages/frontend/components/collection/layouts/index.ts
@@ -1,2 +1,3 @@
 export * from './magazine';
 export * from './card';
+export * from './list';