+++ packages/frontend/components/collection/list/CollectionList.tsx
@@ -17,7 +17,7 @@ import {
   ID,
 } from '@orpington-news/shared';
 import { usePullToRefresh } from '@utils';
-import { CardItem, MagazineItem } from '../layouts';
+import { CardItem, ListItem, MagazineItem } from '../layouts';
 import { RefreshIndicator } from './RefreshIndicator';
 
 export interface CollectionListProps {
@@ -42,6 +42,8 @@ const getListItem = (layout: CollectionLayout) => {
       return MagazineItem;
     case 'card':
       return CardItem;
+    case 'list':
+      return ListItem;
   }
 };
 
@@ -73,6 +75,8 @@ export const CollectionList: React.FC<CollectionListProps & BoxProps> = (
     onRefresh,
   });
 
+  const Skeleton = layout === 'list' ? SkeletonList : SkeletonBox;
+
   if (isLoading) {
     return (
       <VStack
@@ -84,7 +88,7 @@ export const CollectionList: React.FC<CollectionListProps & BoxProps> = (
         px={3}
       >
         {genN(10).map((x) => (
-          <SkeletonBox key={x} />
+          <Skeleton key={x} />
         ))}
       </VStack>
     );
@@ -122,14 +126,12 @@ export const CollectionList: React.FC<CollectionListProps & BoxProps> = (
         itemContent={(index, data) => (
           <Item
             item={data}
-            py={2}
-            pr={3}
             isActive={data.id === activeArticleId}
             data-test={`item-id-${data.id}`}
           />
         )}
         components={{
-          Footer: canFetchMoreItems ? SkeletonBox : undefined,
+          Footer: canFetchMoreItems ? Skeleton : undefined,
         }}
       />
     </Box>
@@ -138,9 +140,19 @@ export const CollectionList: React.FC<CollectionListProps & BoxProps> = (
 
 const SkeletonBox: React.FC = (props) => {
   return (
-    <Box pb={4}>
+    <Box pb={4} pr={4}>
       <SkeletonText mt="4" noOfLines={1} spacing="4" skeletonHeight={4} />
       <SkeletonText mt="4" noOfLines={2} spacing="4" />
+    </Box>
+  );
+};
+
+const SkeletonList: React.FC = (props) => {
+  return (
+    <Box pb={4} pl={1} pr={6}>
+      <SkeletonText mt="4" noOfLines={1} spacing="4" skeletonHeight={4} />
+      <SkeletonText mt="4" noOfLines={1} spacing="4" skeletonHeight={4} />
+      <SkeletonText mt="4" noOfLines={1} spacing="4" skeletonHeight={4} />
     </Box>
   );
 };